+++ index.js
@@ -1,5 +1,22 @@
 const _ = require('lodash')
-const hapi = require('hapi')
+const joi = require('joi')
+const pkg = require('./package.json')
+
+/**
+ * @type {Object}
+ * @private
+ *
+ * @description
+ * Store internal objects
+ */
+const internals = {
+  cb: {
+    serialize: (route) => `${route.method}:${route.path}`
+  },
+  scheme: {
+    labels: joi.array().items(joi.string().min(1)).min(1).single()
+  }
+}
 
 /**
  * @function
@@ -20,17 +37,40 @@ function getRoutes (entry) {
  * @public
  *
  * @description
- * Get flattened list of all defined routes
+ * Plugin to get list of defined routes
  *
  * @param {hapi.Server} server The related hapi server instance
- * @returns {Array.<?Object>} Flattened list of routes
+ * @param {Object} pluginOptions The plugin options
+ * @param {Function} next The callback to continue in the chain of plugins
  */
-function wozu (server) {
-  if (!server || !hapi.Server.prototype.isPrototypeOf(server)) {
-    throw SyntaxError('The passed parameter have to be a `hapi.Server` instance.')
-  }
+function wozu (server, pluginOptions, next) {
+  /**
+   * @function
+   * @public
+   *
+   * @description
+   * Get flattened list of all defined routes
+   *
+   * @param {string | Array.<string>} [labels] Labels to select specific connections
+   * @returns {Array.<?Object>} Flattened list of routes
+   */
+  server.decorate('server', 'wozu', function decorator (labels) {
+    joi.assert(labels, internals.scheme.labels, 'The parameter "labels" is invalid. Its')
 
-  return _.flatten(server.table().map(getRoutes))
+    const connections = labels ? server.select(labels) : server
+    const routeList = _.flatten(connections.table().map(getRoutes))
+    const sorted = _.sortBy(routeList, internals.cb.serialize)
+
+    return _.sortedUniqBy(sorted, internals.cb.serialize)
+  })
+
+  next()
 }
 
-module.exports = wozu
+wozu.attributes = {
+  pkg
+}
+
+module.exports = {
+  register: wozu
+}