+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - **add:** changelog
 - **add:** dependencies
 - **fix:** npm scripts
+- **fix:** transform util into plugin